refactoring + удаление всех сессий при смене пароля

+++ app/routes/auth.py
@@ -8,7 +8,14 @@ from app.db.session import get_db_session
 from app.models.models import User
 from app.schemas.user import UserChangePassword, UserCreate, UserLogin, UserNewPassword, UserRead
 from app.services.auth import hash_paasword, verify_password
-from app.services.jwt import create_token, delete_token, verify_token
+from app.services.jwt import (
+    create_token,
+    decode_token,
+    delete_all_user_access_tokens,
+    delete_token,
+    verify_token,
+)
+from app.services.redis import get_redis_key
 
 auth_router = APIRouter()
 
@@ -38,7 +45,7 @@ async def register(data: UserCreate,
     token = await create_token(user_id=user.id,
                          token_type='email')
 
-    confirm_link = f"http://localhost:8000/auth/confirm-email?token={token}"
+    confirm_link = f"{config.app.url}/auth/confirm-email?token={token}"
 
     return {"user": user, "confirm_link": confirm_link}
 
@@ -80,6 +87,9 @@ async def login(
     if not user or not verify_password(data.password, user.hashed_password):
         raise HTTPException(status_code=401, detail='Invalid credentials')
 
+    if not user or not user.is_active:
+        raise HTTPException(status_code=403, detail="Email not confirmed")
+
     token = await create_token(user.id, token_type='access')
 
     response.set_cookie(
@@ -102,6 +112,9 @@ async def get_me(request: Request):
         raise HTTPException(status_code=401,
                             detail='Unauthoriszed')
 
+    if not user or not user.is_active:
+        raise HTTPException(status_code=403, detail="Email not confirmed")
+
     return user
 
 
@@ -110,8 +123,15 @@ async def logout(request: Request, response: Response):
     token = request.cookies.get("access_token")
 
     if token:
-        redis_key = f"access:{token}"
-        await delete_token(redis_key)
+        payload = await decode_token(token)
+        user_id = payload.get("sub") if payload else None
+
+        if user_id:
+            redis_key = get_redis_key(token=token,
+                                  user_id=user_id,
+                                  token_type='access')
+
+            await delete_token(redis_key)
 
     response = JSONResponse(content={"detail": "Logged out"})
     response.delete_cookie("access_token")
@@ -132,12 +152,20 @@ async def password_reset(
         raise HTTPException(status_code=404,
                             detail='User not found')
 
+    if not user or not user.is_active:
+        raise HTTPException(status_code=403, detail="Email not confirmed")
+
     token = await create_token(user_id=user.id,
                                token_type='password')
 
-    reset_link = f"http://localhost:8000/auth/reset-password?token={token}"
+    reset_link = f"{config.app.url}/auth/redirect?token={token}"
 
-    return {"detail": reset_link}
+    return {"reset_link": reset_link}
+
+
+@auth_router.get('/redirect')
+async def get_token_for_new_password(token: str = Query(...)):
+    return {"token": token}
 
 
 @auth_router.post('/new_password')
@@ -160,6 +188,8 @@ async def new_password(
                             detail='User not found')
 
     user.hashed_password = hash_paasword(data.new_password)
-    await db.commit()
 
-    return {"detail": "Password updated successfully ✅"}
+    await delete_all_user_access_tokens(user_id)
+
+    await db.commit()
+    return {"detail": "Password changed successfully ✅"}
